LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-no-cr-expected.txt

Handshake should fail when the first line does not end with CRLF.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connected is false
PASS wsOrigin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
